### 2024-12-05 26475b5
Removed: sections Introduction, and Architecture; Added: short ref. to preprint for extended introduction
