+++ paper.bib
@@ -1,17 +1,3 @@
-@misc{machlab_llm_2024,
-	title = {{LLM} {In}-{Context} {Recall} is {Prompt} {Dependent}},
-	url = {http://arxiv.org/abs/2404.08865},
-	abstract = {The proliferation of Large Language Models (LLMs) highlights the critical importance of conducting thorough evaluations to discern their comparative advantages, limitations, and optimal use cases. Particularly important is assessing their capacity to accurately retrieve information included in a given prompt. A model’s ability to do this significantly influences how effectively it can utilize contextual details, thus impacting its practical efficacy and dependability in real-world applications. Our research analyzes the in-context recall performance of various LLMs using the needle-in-a-haystack method. In this approach, a factoid (the “needle”) is embedded within a block of filler text (the “haystack”), which the model is asked to retrieve. We assess the recall performance of each model across various haystack lengths and with varying needle placements to identify performance patterns. This study demonstrates that an LLM’s recall capability is not only contingent upon the prompt’s content but also may be compromised by biases in its training data. Conversely, adjustments to model architecture, training strategy, or fine-tuning can improve performance. Our analysis provides insight into LLM behavior, offering direction for the development of more effective applications of LLMs.},
-	language = {en},
-	urldate = {2024-04-26},
-	publisher = {arXiv},
-	author = {Machlab, Daniel and Battle, Rick},
-	month = apr,
-	year = {2024},
-	note = {arXiv:2404.08865 [cs]},
-	keywords = {Computer Science - Computation and Language, Computer Science - Machine Learning},
-}
-
 @article{page_prisma_2021,
 	title = {The {PRISMA} 2020 statement: an updated guideline for reporting systematic reviews},
 	volume = {10},
@@ -130,44 +116,6 @@ Section: Communication},
 	pages = {e072254},
 }
 
-@article{hagendorff_human-like_2023,
-	title = {Human-like intuitive behavior and reasoning biases emerged in large language models but disappeared in {ChatGPT}},
-	volume = {3},
-	issn = {2662-8457},
-	url = {https://www.nature.com/articles/s43588-023-00527-x},
-	doi = {10.1038/s43588-023-00527-x},
-	abstract = {Abstract
-            We design a battery of semantic illusions and cognitive reflection tests, aimed to elicit intuitive yet erroneous responses. We administer these tasks, traditionally used to study reasoning and decision-making in humans, to OpenAI’s generative pre-trained transformer model family. The results show that as the models expand in size and linguistic proficiency they increasingly display human-like intuitive system 1 thinking and associated cognitive errors. This pattern shifts notably with the introduction of ChatGPT models, which tend to respond correctly, avoiding the traps embedded in the tasks. Both ChatGPT-3.5 and 4 utilize the input–output context window to engage in chain-of-thought reasoning, reminiscent of how people use notepads to support their system 2 thinking. Yet, they remain accurate even when prevented from engaging in chain-of-thought reasoning, indicating that their system-1-like next-word generation processes are more accurate than those of older models. Our findings highlight the value of applying psychological methodologies to study large language models, as this can uncover previously undetected emergent characteristics.},
-	language = {en},
-	number = {10},
-	urldate = {2024-05-28},
-	journal = {Nature Computational Science},
[…]
