+++ paper.md
@@ -26,31 +26,7 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Systematic literature reviews play a crucial role in synthesizing research findings across various disciplines. However, traditional approaches are typically labor-intensive, time-consuming, and prone to subjectivity, making it challenging to manage large datasets while ensuring consistency and reproducibility. `prismAId` addresses these challenges by automating data extraction and analysis through advanced AI models, significantly reducing the time and effort required. Its compatibility with multiple programming languages and operating systems ensures broad accessibility, while integration with established review protocols like PRISMA 2020 guarantees methodological rigor. `prismAId` provides a user-friendly interface that requires no coding skills, enabling researchers from diverse backgrounds to perform comprehensive literature analyses efficiently.
-
-# Introduction
-
-A systematic literature review (SLR) is a structured method of synthesizing existing research to answer specific questions. Unlike traditional literature reviews, SLRs employ rigorous and transparent procedures to minimize bias and ensure reproducibility. This approach involves clearly defining research questions, systematically searching for relevant studies, critically appraising their quality, and synthesizing findings to provide comprehensive insights into a particular topic.
-
-The importance of SLRs lies in their ability to consolidate vast amounts of information, offering evidence-based conclusions that inform practice, policy, and future research. By evaluating and integrating findings from multiple studies, SLRs help identify consistent patterns, resolve discrepancies, and highlight areas needing further investigation. This process not only advances scientific knowledge but also supports decision-making across various fields.
-
-Many protocols exist for conducting SLRs [@schiavo_prospero_2019], including PRISMA 2020 [@page_prisma_2021], which inspired the name of our tool, as well as preregistration services and templates designed to standardize and support SLR processes.
-
-Many recent contributions discuss the role that AI models can play in supporting SLRs. The integration of AI is predominantly highlighted in the screening phase, where machine learning algorithms are employed to filter and prioritize studies, significantly reducing the burden of manual work. However, the broader application of AI across the entire review process, including data extraction and risk of bias assessment, reveals both potential benefits and significant limitations.
-
-The review by @blaizot_using_2022 provides an in-depth analysis of AI methods in SLRs, particularly in health sciences. The authors identify title and abstract screening as the primary area of application, where 73% of AI tools are concentrated. Machine learning algorithms assist in filtering studies, expediting the initial triage process. Despite the efficiency gains, the authors caution that current AI tools require substantial human input for validation, as fully autonomous screening remains limited. In data extraction, tools like RobotReviewer show promise but exhibit frequent errors, necessitating manual corrections. AI methods also attempt to automate risk of bias assessment, yet significant issues with accuracy and consistency demand human oversight. The authors conclude that while AI can enhance efficiency in certain stages of SLRs, challenges related to data quality, model transparency, and the need for expert intervention persist.
-
-@van_de_schoot_open_2021 propose AI models, especially those using active learning, as tools to streamline the screening phase. These models prioritize relevant records based on human feedback, reducing the number of studies needing manual evaluation. However, AI models often struggle with complex and evolving inclusion criteria, especially when relevance depends on nuanced context. The selection phase also benefits from AI, as machine learning tools help predict the eligibility of full-text articles. However, reliance on simple text features may result in missed insights, making AI effective for initial triage but insufficient for detailed assessments.
-
[…]
