+++ paper.md
@@ -75,17 +75,11 @@ The standard review prompt in `prismAId` may consist of six main blocks, which a
 
 The six blocks include: the persona, which defines the model’s role to provide context for its responses; the task, specifying the exact information extraction task; and the expected result, typically a structured JSON object formatted according to the configuration file. The failsafe block offers guidance for the model on how to handle uncertain situations (e.g., leave the response empty if unsure). The definitions block includes detailed explanations for the keys and possible values of the JSON object, minimizing the risk of misinterpretation. Finally, the examples block includes sample manuscripts and their expected outputs, offering concrete instances that help the model understand the desired format and content of the extracted information, thus reducing the risk of misinterpretation further.
 
-# Discussion and Conclusions
+# Conclusions
 
-`prismAId`, for which @boero_extended_2024 provides details on the software architecture and a comparison with existing AI approaches to SLRs, has the potential to make a significant impact on the research community. By leveraging advanced AI capabilities, the tool offera a faster, more efficient, and scalable alternative to traditional methods to SLRs. The focus on open science principles enhances transparency and reproducibility, making systematic reviews more accessible and easier to update. This shift is especially important in fields where the volume of literature is growing rapidly, helping researchers keep up with the latest developments while reducing the manual workload.
-
-Despite its strengths, there are potential limitations that need to be addressed. The reliance on generative AI models means that the quality of extracted information depends on the model's understanding and performance, which can vary across different contexts and disciplines. Ensuring that prompts are well-designed and tailored to specific use cases is crucial, as poorly constructed prompts could lead to inaccuracies or misinterpretations. Additionally, while `prismAId` integrates with leading AI models and provides robust features, further testing and feedback are needed to optimize its performance across diverse research areas. Planned future developments include expanded cloud support, integration with new AI models, and enhancements to the user interface to streamline configuration and improve usability. 
-
-An important area for future development in the coming years will be the creation of standardized approaches to assess potential biases, both in the literature being reviewed and within the review configuration itself. By establishing clear guidelines and robust evaluation methods, `prismAId` aims to address the risks of bias that may arise from the selection of manuscripts, the framing of prompts, and the interpretation of AI-generated outputs. These advancements will contribute to enhancing the reliability and credibility of SLRs, further aligning with the principles of open science and evidence-based research.
+`prismAId`, for which @boero_extended_2024 provides details on the software architecture and a comparison with existing AI approaches to SLRs, has the potential to make a significant impact on the research community. By leveraging advanced AI capabilities, the tool offers a faster, more efficient, and scalable alternative to traditional methods to SLRs. The focus on open science principles enhances transparency and reproducibility, making systematic reviews more accessible and easier to update. This shift is especially important in fields where the volume of literature is growing rapidly, helping researchers keep up with the latest developments while reducing the manual workload.
 
 Anecdotal evidence suggests that `prismAId` outperforms traditional SLR methods in both accuracy and efficiency. Preliminary applications, such as the one described by @boero_ai-enhanced_2024, demonstrate significant time savings, with reviews completed in a fraction of the time required for manual analysis. The cost of running these smaller-scale reviews was minimal and has been continuously and significantly decreasing, reflecting the rapid advancements in the availability and affordability of AI model services.
-
-In summary, `prismAId` represents a major step forward in the field of SLRs, advancing the principles of open science by enabling faster, more transparent, and replicable analyses. Its key contributions lie in the automation of data extraction, the standardization of prompts, and the provision of accessible, multi-platform tools for researchers. Its open-source nature is fundamental to `prismAId`, as it enables thorough code inspection, facilitates the replication of results, and invites the broader research community to actively contribute to its ongoing development and improvement.
 
 # Acknowledgements
 
