+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'An Introduction to `prismAId`: Open-Source and Open Science AI for Advancing Information Extraction in Systematic Reviews'
+title: 'An Introduction to prismAId: Open-Source and Open Science AI for Advancing Information Extraction in Systematic Reviews'
 tags:
   - systematic literature review
   - generative AI
@@ -29,7 +29,6 @@ bibliography: paper.bib
 Systematic literature reviews (SLRs) play a crucial role in synthesizing research findings across various disciplines. However, traditional approaches are typically labor-intensive, time-consuming, and prone to subjectivity, making it challenging to manage large datasets while ensuring consistency and reproducibility. `prismAId` addresses these challenges by automating data extraction and analysis through advanced AI models, significantly reducing the time and effort required. Its compatibility with multiple programming languages and operating systems ensures broad accessibility, while integration with established review protocols like PRISMA 2020 guarantees methodological rigor. `prismAId` provides a user-friendly interface that requires no coding skills, enabling researchers from diverse backgrounds to perform comprehensive literature analyses efficiently.
 
 # Features and Capabilities
-
 `prismAId` is compatible with any systematic review protocol [@schiavo_prospero_2019], including the widely adopted PRISMA 2020 guidelines [@page_prisma_2021]. It focuses specifically on the analysis phase, which is a key concluding component of all systematic review protocols, alongside the reporting stage.
 
 ![Traditional approach to literature analysis vs. `prismAId`-supported review workflow.\label{fig:workflow}](workflow.png){width="50%"}
@@ -44,15 +43,11 @@ Building on the tool's aim, `prismAId` includes the capability to preprocess and
 
 Additionally, `prismAId` integrates with Zotero, a popular open source reference management software, allowing users to access shared collections and groups of scientific literature seamlessly. This integration simplifies document retrieval and enhances collaboration. The connection to Zotero is automatic once the API access is configured, as detailed in the `prismAId` documentation. This feature streamlines the workflow, making it easier for users to manage and analyze large sets of scientific manuscripts.
 
-`prismAId` integrates with various large language models (LLMs) from the four main AI providers: OpenAI, Google Cloud AI, Anthropic, and Cohere. This includes access to state-of-the-art models such as many variants of GPT, Gemini, Claude, and Command. The integration is handled through native APIs, allowing for seamless access to advanced generative AI capabilities. Depending on project needs and available opportunities, support for cloud-based deployments of these models will be expanded. This variety ensures that users can leverage the most up-to-date technology while benefiting from different pricing options, making `prismAId` adaptable to a range of project budgets and requirements.
-
-The main results dataset generated by `prismAId` can be exported in two formats: a tabular CSV and a nested JSON structure. These options ensure compatibility with a wide range of analysis tools and software. The CSV format is ideal for spreadsheet-based analysis and integration with statistical packages, while the JSON format supports more complex data structures and is suitable for use in web applications or advanced data processing pipelines. Together, these formats cover all potential use cases and enable seamless data import and interoperability.
+`prismAId` integrates with various large language models (LLMs) from the four main AI providers: OpenAI, Google Cloud AI, Anthropic, and Cohere. This includes access to state-of-the-art models such as many variants of GPT, Gemini, Claude, and Command. The integration is handled through native APIs, allowing for seamless access to advanced generative AI capabilities. Depending on project needs and available opportunities, support for cloud-based deployments of these models will be expanded. The main results dataset generated by `prismAId` can be exported in two formats: a tabular CSV and a nested JSON structure. These options ensure compatibility with a wide range of analysis tools and software.
 
[…]
