+++ paper.bib
@@ -178,7 +178,6 @@ Section: Communication},
 	month = aug,
 	year = {2024},
 	pages = {116},
-	file = {PDF:/var/home/ribo/Zotero/storage/EMFAJHKW/Boero - 2024 - An AI-Enhanced Systematic Review of Climate Adaptation Costs Approaches and Advancements, 2010–2021.pdf:application/pdf},
 }
 
 @misc{boero_extended_2024,
@@ -189,10 +188,9 @@ Section: Communication},
 	abstract = {`prismAId` is an open-source tool designed to streamline systematic literature reviews by leveraging generative AI models for information extraction. It offers an accessible, efficient, and replicable method for extracting and analyzing data from scientific literature, eliminating the need for coding expertise. Supporting various review protocols, including PRISMA 2020, `prismAId` is distributed across multiple platforms — Go, Python, Julia, R — and provides user-friendly binaries compatible with Windows, macOS, and Linux. The tool integrates with leading large language models (LLMs) such as OpenAI's GPT series, Google's Gemini, Cohere's Command, and Anthropic's Claude, ensuring comprehensive and up-to-date literature analysis. `prismAId` facilitates systematic reviews, enabling researchers to conduct thorough, fast, and reproducible analyses, thereby advancing open science initiatives.},
 	language = {en-us},
 	urldate = {2024-12-05},
-	publisher = {OSF},
+	publisher = {MetaArXiv},
 	author = {Boero, Riccardo},
 	month = nov,
 	year = {2024},
 	keywords = {Generative AI, Large Language Model, Open Source, Sopen Science, Systematic Literature Review},
-	file = {OSF Preprint:/var/home/ribo/Zotero/storage/52HDSZ27/Boero - 2024 - An Extended Introduction to the `prismAId` Tool Open-Source and Open Science AI for Systematic Revi.pdf:application/pdf},
 }
