### 2024-12-05 #7569 @ricboer0
Thanks @crvernon for the support so far.
A couple of questions.
A) I managed to cut half of the original submission, now at 1,800 words. Is it good enough? Going down now is more difficult.
B) is there a way to reference the github repository from the markdown? I put it as an entry in references but i see a nice link on the left side of the article proof. Otherwise I'll just mention the repository in text with no link.
