### 2024-12-05 ca6adb5
Removed: unused references & to Github repo
