+++ paper.md
@@ -50,7 +50,7 @@ Additionally, `prismAId` integrates with Zotero, a popular open source reference
 Moreover, `prismAId` supports ensemble reviews, a powerful feature that allows the same review to be run across different models, both within a single provider and across multiple providers. This approach enables the calculation of uncertainty estimates for the extracted answers, helping to assess the robustness of the results and providing a more comprehensive analysis.
 
 # User Experience
-The primary access to `prismAId` is through standalone binaries, compiled for Windows, macOS, and Linux, with support for both AMD64 (x86-64) and ARM64 architectures. These self-contained binaries allow users to run and manage review projects, as well as initialize project configuration files directly from the terminal. They do not require the installation of any additional software beyond terminal access, making them simple and convenient to use. The binaries can be easily downloaded from the release page on the project GitHub repository [@prismaid].
+The primary access to `prismAId` is through standalone binaries, compiled for Windows, macOS, and Linux, with support for both AMD64 (x86-64) and ARM64 architectures. These self-contained binaries allow users to run and manage review projects, as well as initialize project configuration files directly from the terminal. They do not require the installation of any additional software beyond terminal access, making them simple and convenient to use. The binaries can be easily downloaded from the release page on the project GitHub repository.
 
 For users with more technical expertise, there are additional programmatic options for accessing `prismAId`. One option is to integrate the tool directly by using the Go module available on the GitHub repo. Alternatively, `prismAId` can be installed as a Python package via PyPI, as a Julia package from the General package registry, or as an R package through R-universe. These options offer more flexibility and are suitable for integrating `prismAId` into custom workflows and advanced scripting environments.
 
