+++ paper.bib
@@ -14,108 +14,6 @@
 	pages = {89},
 }
 
-@article{blaizot_using_2022,
-	title = {Using artificial intelligence methods for systematic review in health sciences: {A} systematic review},
-	volume = {13},
-	issn = {1759-2887},
-	shorttitle = {Using artificial intelligence methods for systematic review in health sciences},
-	doi = {10.1002/jrsm.1553},
-	abstract = {The exponential increase in published articles makes a thorough and expedient review of literature increasingly challenging. This review delineated automated tools and platforms that employ artificial intelligence (AI) approaches and evaluated the reported benefits and challenges in using such methods. A search was conducted in 4 databases (Medline, Embase, CDSR, and Epistemonikos) up to April 2021 for systematic reviews and other related reviews implementing AI methods. To be included, the review must use any form of AI method, including machine learning, deep learning, neural network, or any other applications used to enable the full or semi-autonomous performance of one or more stages in the development of evidence synthesis. Twelve reviews were included, using nine different tools to implement 15 different AI methods. Eleven methods were used in the screening stages of the review (73\%). The rest were divided: two in data extraction (13\%) and two in risk of bias assessment (13\%). The ambiguous benefits of the data extractions, combined with the reported advantages from 10 reviews, indicating that AI platforms have taken hold with varying success in evidence synthesis. However, the results are qualified by the reliance on the self-reporting of the review authors. Extensive human validation still appears required at this stage in implementing AI methods, though further evaluation is required to define the overall contribution of such platforms in enhancing efficiency and quality in evidence synthesis.},
-	language = {eng},
-	number = {3},
-	journal = {Research Synthesis Methods},
-	author = {Blaizot, Aymeric and Veettil, Sajesh K. and Saidoung, Pantakarn and Moreno-Garcia, Carlos Francisco and Wiratunga, Nirmalie and Aceves-Martins, Magaly and Lai, Nai Ming and Chaiyakunapruk, Nathorn},
-	month = may,
-	year = {2022},
-	pmid = {35174972},
-	keywords = {artificial intelligence, Artificial Intelligence, evidence synthesis, Humans, machine learning, Machine Learning, Medicine, systematic reviews, Systematic Reviews as Topic},
-	pages = {353--362},
-}
-
-@article{fabiano_how_nodate,
-	title = {How to optimize the systematic review process using {AI} tools},
-	volume = {n/a},
-	copyright = {© 2024 The Authors. JCPP Advances published by John Wiley \& Sons Ltd on behalf of Association for Child and Adolescent Mental Health.},
-	issn = {2692-9384},
-	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcv2.12234},
-	doi = {10.1002/jcv2.12234},
[…]
