### 2024-12-05 94add10
Fixed: figures size
