### 2024-12-05 #7569 @philip928lin
> 👋 @philip928lin - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@crvernon Yes, I am willing to review this submission. Thanks for the invitation.
