### 2024-12-24 #7569 @jhculb
Hi, I think I should have some time to do so, yes :)

> 👋 @jhculb - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
