### 2024-12-24 #7616 @editorialbot
[REVIEW]: An Introduction to `prismAId`: Open-Source and Open Science AI for Advancing Information Extraction in Systematic Reviews
**Submitting author:** @ricboer0 (<a href="http://orcid.org/0000-0002-7468-9096">Riccardo Boero</a>)
**Repository:** https://github.com/open-and-sustainable/prismaid
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** joss
