### 2024-12-28 #7616 @jhculb
Hi @ricboer0,

I'm getting a 404 on the github.io documentation link in the readme: https://open-and-sustainable.github.io/prismAId/

Is this just me?
