### 2024-12-28 #7616 @jhculb
Hi @ricboer0,

Your paper may be improved with a brief survey of past and current academic information extraction tools, as at the moment the state of the field section of the paper feels lacking.

Perhaps LISC https://github.com/lisc-tools/lisc ?
