### 2024-12-28 #7616 @philip928lin
> Hi @ricboer0,
>
> I'm getting a 404 on the github.io documentation link in the readme: https://open-and-sustainable.github.io/prismAId/
>
> Is this just me? @philip928lin

@jhculb I encountered same issue as you. Multiple hyperlinks in the readme have same issue. @ricboer0 please double check the links.
