### 2024-12-28 #7616 @philip928lin
Hi @ricboer0,
It is nice to see the online Review Configurator. However, the outputted configuration file has incorrect value format that leads to error when running the code. Please see below.

The online configurator did not output the correct format

_Error loading project configuration: toml: line 27 (last key "project.llm.1.temperature"): incompatible types: TOML value has type string; destination has type float
Error: toml: line 27 (last key "project.llm.1.temperature"): incompatible types: TOML value has type string; destination has type float_

Namely, the following lines in TOML file
```
temperature = "0.01"
tpm_limit = "0"
rpm_limit = "0"
```

should be

```
temperature = 0.01
tpm_limit = 0
rpm_limit = 0
```
