### 2024-12-28 #7616 @philip928lin
Hi @ricboer0,

I also tried to play with the Zotero Integration. I appreciate this feature. However, more detailed steps and expected outputs can be provided in the documentation to ease usage.

1. A minor correction in the documentation: "Privacy Tab" should be "Security Tab"
![Pasted image 20241227115411](https://github.com/user-attachments/assets/83fddbee-be24-4f57-b625-fa7376c79d33)

2. Adding the example Zotero configuration block to the documentation would be helpful. E.g.,

```
[project.zotero]
user = ""
api_key = ""
group = ""
 ```

3. I could not see the review results when running with Zotero integration.
I tried to follow the instruction to run the literature review using Zotero integration. After running, I can see the "zotero" folder is created with the extracted texts as described in the documentation, but I have a hard time finding the review output like "output_test.csv" or "lit_test_summary.txt" at the given output path. No error messages were shown. I was using a Python wrapper and called the python script through the command line. @jhculb, do you have the same issue here?

@ricboer0, I would appreciate more detailed instructions and examples on the Zotero integration feature.

Also, another side note, when I executed the Python script in the IDE like Spyder, it restarted a new kernel in the middle of the process every time. However, this may be due to the IDE's internal issue. I can successfully run the Python script when directly called from the command line.
