### 2024-12-30 open-and-sustainable/prismaid#107 @ricboer0
Fix links to documentation website in README.md
https://github.com/openjournals/joss-reviews/issues/7616#issuecomment-2564380307
