### 2024-12-30 #7616 @ricboer0
> > Hi @ricboer0,
> > I'm getting a 404 on the github.io documentation link in the readme: https://open-and-sustainable.github.io/prismAId/
> > Is this just me? @philip928lin
>
> @jhculb I encountered same issue as you. Multiple hyperlinks in the readme have same issue. @ricboer0 please double check the links.

Hi @jhculb and @philip928lin,
thank you so much for all the useful comments and suggestions.

I am working to double check and fix links to documentation website
https://github.com/open-and-sustainable/prismaid/issues/107
I will let you know when fixes are ready and merged.
