### 2024-12-30 #7616 @ricboer0
> Hi @ricboer0,
>
> Your paper may be improved with a brief survey of past and current academic information extraction tools, as at the moment the state of the field section of the paper feels lacking.
>
> Perhaps LISC https://github.com/lisc-tools/lisc ? Or AutoIE: https://doi.org/10.1007/978-981-97-5495-3_32 ?

Hi @jhculb,
I did include a review of AI tools and approaches to SLRs in the first submission of the paper to JOSS. It was actually a review of reviews on the topic because it is already a vast field in the literature.

However, following guidelines from the editor @crvernon I significantly reduced the length of the paper and removed that section (among others). You can read that 'introduction' section in the preprint that corresponds to the first submission of this paper at [https://doi.org/10.31222/osf.io/wh8qn](https://doi.org/10.31222/osf.io/wh8qn).

I am open to any suggestion. In this shorter version of the paper I already reference the extended introduction accessible in the preprint for a more complete description of alternative approaches but if this is not enough I would be happy to bring that removed section back into the paper. It is just a matter of space constraints, if they matter.
