### 2024-12-30 open-and-sustainable/prismaid#108 @ricboer0
Fix formatting issues in online Review Configurator
https://github.com/openjournals/joss-reviews/issues/7616#issuecomment-2564381569
