### 2024-12-30 #7616 @ricboer0
> Hi @ricboer0, It is nice to see the online Review Configurator. However, the outputted configuration file has incorrect value format that leads to error when running the code. Please see below.
>
> The online configurator did not output the correct format
>
> _Error loading project configuration: toml: line 27 (last key "project.llm.1.temperature"): incompatible types: TOML value has type string; destination has type float Error: toml: line 27 (last key "project.llm.1.temperature"): incompatible types: TOML value has type string; destination has type float_
>
> Namely, the following lines in TOML file
>
> ```
> temperature = "0.01"
> tpm_limit = "0"
> rpm_limit = "0"
> ```
>
> should be
>
> ```
> temperature = 0.01
> tpm_limit = 0
> rpm_limit = 0
> ```
>
> Also, if the online Review Configurator can automatically format the path string for users, it can eliminate the potential errors as shown below.
>
> ```
> Error loading project configuration: toml: line 9 (last key "project.configuration.results_file_name"): expected eight hexadecimal digits after '\U', but got "C:\\Us" instead
> Error: toml: line 9 (last key "project.configuration.results_file_name"): expected eight hexadecimal digits after '\U', but got "C:\\Us" instead
> ```

Hi @philip928lin,
thanks for finding the bug and the suggestion on path formatting. I will work on these issues at https://github.com/open-and-sustainable/prismaid/issues/108 and let you know once finished.
