### 2024-12-30 open-and-sustainable/prismaid#109 @ricboer0
Improve documentation on Zotero integration
https://github.com/openjournals/joss-reviews/issues/7616#issuecomment-2564386652
