### 2024-12-30 #7616 @ricboer0
> Hi @ricboer0,
>
> I also tried to play with the Zotero Integration. I appreciate this feature. However, more detailed steps and expected outputs can be provided in the documentation to ease usage.
>
>     1. A minor correction in the documentation: "Privacy Tab" should be "Security Tab"
>        ![Pasted image 20241227115411](https://private-user-images.githubusercontent.com/23210992/399098495-83fddbee-be24-4f57-b625-fa7376c79d33.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzU1NDY0MjgsIm5iZiI6MTczNTU0NjEyOCwicGF0aCI6Ii8yMzIxMDk5Mi8zOTkwOTg0OTUtODNmZGRiZWUtYmUyNC00ZjU3LWI2MjUtZmE3Mzc2Yzc5ZDMzLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDEyMzAlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQxMjMwVDA4MDg0OFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWM2NGFhN2Y1NjFhZGVlYTNkOWQ0M2JjMTExMzhjMTkwOGViYmRjZTE4ODE3MTI5YmIwM2E0ZDI3NWVhM2M4NDEmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.PHZ7mkm9t7lcpylUCtKYbAdr9NrcS-LcjU_X3N1qypI)
>
>     2. Adding the example Zotero configuration block to the documentation would be helpful. E.g.,
>
>
> ```
> [project.zotero]
> user = ""
> api_key = ""
> group = ""
> ```
>
>     3. I could not see the review results when running with Zotero integration.
>        I tried to follow the instruction to run the literature review using Zotero integration. After running, I can see the "zotero" folder is created with the extracted texts as described in the documentation, but I have a hard time finding the review output like "output_test.csv" or "lit_test_summary.txt" at the given output path. No error messages were shown. I was using a Python wrapper and called the python script through the command line. @jhculb, do you have the same issue here?
>
>
> @ricboer0, I would appreciate more detailed instructions and examples on the Zotero integration feature.
>
> Also, another side note, when I executed the Python script in the IDE like Spyder, it restarted a new kernel in the middle of the process every time. However, this may be due to the IDE's internal issue. I can successfully run the Python script when directly called from the command line.

Hi @philip928lin,
On issues .1 and .2 I will improve the documentation in https://github.com/open-and-sustainable/prismaid/issues/109

On issue .3: since you can find the Zotero manuscripts on your machine, it means that the Zotero integration is properly configured and actually working: it connects to the API, it finds the collection, downloads the pdfs and 'tries' the conversion to .txt.
Though I don't know anything about Spyder, it seems like you are encountering an error later in the execution of the review logic or in the phase saving results.
Please set the logging level to medium or high: through log inspection you should be able to spot the problem quite easily.
