### 2024-12-30 ad01ac3
Fix links in README on project root and Julia package
