### 2024-12-30 84365f0
Fix online Review Configurator, minor typo in README
