### 2024-12-30 286763f
Add details on Zotero integration in documentation site
