### 2024-12-30 #7616 @ricboer0
Hi @jhculb and @philip928lin,
I have completed and merged all the issues I opened to incorporate your suggestions, and I have also rebased the JOSS submission branch.

Please verify if the changes address the comments you raised.

Additionally, could you let me know if I can extend the paper to include a review of the state of the field? This would help to clearly highlight the novelty of this tool, but I believe it cannot be added at the expense of other sections in the paper, as they are all at least as essential.
