### 2024-12-30 #7616 @jhculb
> > Hi @ricboer0,
> >
> > Your paper may be improved with a brief survey of past and current academic information extraction tools, as at the moment the state of the field section of the paper feels lacking.
> >
> > Perhaps LISC https://github.com/lisc-tools/lisc ? Or AutoIE: https://doi.org/10.1007/978-981-97-5495-3_32 ?
>
> Hi @jhculb,
> I did include a review of AI tools and approaches to SLRs in the first submission of the paper to JOSS. It was actually a review of reviews on the topic because it is already a vast field in the literature.
>
> However, following guidelines from the editor @crvernon I significantly reduced the length of the paper and removed that section (among others). You can read that 'introduction' section in the preprint that corresponds to the first submission of this paper at [https://doi.org/10.31222/osf.io/wh8qn](https://doi.org/10.31222/osf.io/wh8qn).
>
> I am open to any suggestion. In this shorter version of the paper I already reference the extended introduction accessible in the preprint for a more complete description of alternative approaches but if this is not enough I would be happy to bring that removed section back into the paper. It is just a matter of space constraints, if they matter.

Hi @ricboer0,

I will defer to the editor on this (@crvernon - your thoughts?), but I understand that the focus for JOSS' State of the field section is the availability of similar software and the distinction to your tool.

From the review checklist; "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

Having a quick look at your preprint I see a interesting review of the field but no explicit software cited, which may be the reason @crvernon asked for it to be condensed.

I personally would recommend a short review of popular or distubguished software packages / applications providing similar functionality be included. Otherwise a fine paper I believe :)
