### 2024-12-30 #7616 @crvernon
Concerning the length of the paper and the statement of need:

> I personally would recommend a short review of popular or distinguished software packages / applications providing similar functionality be included.

I think this is a reasonable request.  In JOSS we want to keep the paper as succinct as possible and encourage the procedural and in-depth descriptive content to be placed in the software's documentation for lasting benefit that will hopefully be improved as the software evolves.  This makes the JOSS paper for powerful too, since it will not be filled with content that may be quickly outdated as versions of the software change - instead, with a focus on why this software is needed and why it is being introduced as an open-source community research software.

On this note, it is good to cite what is comparable and the value that you believe that your software is adding.  JOSS does not require uniqueness, but the hope is that this transparency in the statement of need section will give users context to quickly decide if the proposed software is one that may benefit their own research.

Thanks!
