### 2025-01-02 1ac57bb
add 'State of the Field' section as requested in review, https://github.com/openjournals/joss-reviews/issues/7616#issuecomment-2565612368
