+++ paper.bib
@@ -14,32 +14,6 @@
 	pages = {89},
 }
 
-@article{schiavo_prospero_2019,
-	title = {{PROSPERO}: {An} {International} {Register} of {Systematic} {Review} {Protocols}},
-	volume = {38},
-	issn = {0276-3869, 1540-9597},
-	shorttitle = {{PROSPERO}},
-	url = {https://www.tandfonline.com/doi/full/10.1080/02763869.2019.1588072},
-	doi = {10.1080/02763869.2019.1588072},
-	abstract = {PROSPERO is an international database of systematic review protocols produced by the University of York’s Center for Research and Dissemination and funded by the National Institute for Health Research. It contains protocols of systematic reviews on health and social care, welfare, public health, education, crime, justice, and health-related international development. PROSPERO compiles a comprehensive listing of systematic review protocols in an attempt to avoid duplication of effort, reduce reporting bias, and promote transparency.},
-	language = {en},
-	number = {2},
-	urldate = {2024-11-10},
-	journal = {Medical Reference Services Quarterly},
-	author = {Schiavo, Julie H.},
-	month = apr,
-	year = {2019},
-	pages = {171--180},
-}
-
-@misc{prismaid-doc,
-  author = {Boero, Riccardo},
-  title = {{prismAId - Documentation website}},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub pages},
-  url = {https://open-and-sustainable.github.io/prismaid/}
-}
 
 @article{boero_ai-enhanced_2024,
 	title = {An {AI}-{Enhanced} {Systematic} {Review} of {Climate} {Adaptation} {Costs}: {Approaches} and {Advancements}, 2010–2021},
@@ -60,6 +34,24 @@
 	pages = {116},
 }
 
+@article{schiavo_prospero_2019,
+	title = {{PROSPERO}: {An} {International} {Register} of {Systematic} {Review} {Protocols}},
+	volume = {38},
+	issn = {0276-3869, 1540-9597},
+	shorttitle = {{PROSPERO}},
+	url = {https://www.tandfonline.com/doi/full/10.1080/02763869.2019.1588072},
+	doi = {10.1080/02763869.2019.1588072},
+	abstract = {PROSPERO is an international database of systematic review protocols produced by the University of York’s Center for Research and Dissemination and funded by the National Institute for Health Research. It contains protocols of systematic reviews on health and social care, welfare, public health, education, crime, justice, and health-related international development. PROSPERO compiles a comprehensive listing of systematic review protocols in an attempt to avoid duplication of effort, reduce reporting bias, and promote transparency.},
+	language = {en},
+	number = {2},
+	urldate = {2024-11-10},
+	journal = {Medical Reference Services Quarterly},
+	author = {Schiavo, Julie H.},
+	month = apr,
+	year = {2019},
+	pages = {171--180},
+}
+
 @misc{boero_extended_2024,
 	title = {An {Extended} {Introduction} to the `{prismAId}` {Tool}: {Open}-{Source} and {Open} {Science} {AI} for {Systematic} {Reviews}},
 	shorttitle = {An {Extended} {Introduction} to the `{prismAId}` {Tool}},
@@ -70,7 +62,246 @@
 	urldate = {2024-12-05},
 	publisher = {MetaArXiv},
 	author = {Boero, Riccardo},
-	month = {Nov},
+	month = nov,
 	year = {2024},
-	keywords = {Generative AI, Large Language Model, Open Source, Sopen Science, Systematic Literature Review},
+	keywords = {Generative AI, Large Language Model, Open Source, Systematic Literature Review, Open Science},
+}
[…]
