### 2025-01-02 4fa3cb2
add 'State of the Field' section as requested in review, https://github.com/openjournals/joss-reviews/issues/7616#issuecomment-2565612368
