+++ paper.md
@@ -21,12 +21,13 @@ bibliography: paper.bib
 ---
 
 # Summary
-
 `prismAId` is an open-source tool designed to streamline systematic literature reviews by leveraging generative AI models for information extraction. It offers an accessible, efficient, and replicable method for extracting and analyzing data from scientific literature, eliminating the need for coding expertise. Supporting various review protocols, including PRISMA 2020, `prismAId` is distributed across multiple platforms -- Go, Python, Julia, R -- and provides user-friendly binaries compatible with Windows, macOS, and Linux. The tool integrates with leading large language models (LLMs) such as OpenAI's GPT series, Google's Gemini, Cohere's Command, and Anthropic's Claude, ensuring comprehensive and up-to-date literature analysis. `prismAId` facilitates systematic reviews, enabling researchers to conduct thorough, fast, and reproducible analyses, thereby advancing open science initiatives.
 
 # Statement of Need
-
 Systematic literature reviews (SLRs) play a crucial role in synthesizing research findings across various disciplines. However, traditional approaches are typically labor-intensive, time-consuming, and prone to subjectivity, making it challenging to manage large datasets while ensuring consistency and reproducibility. `prismAId` addresses these challenges by automating data extraction and analysis through advanced AI models, significantly reducing the time and effort required. Its compatibility with multiple programming languages and operating systems ensures broad accessibility, while integration with established review protocols like PRISMA 2020 guarantees methodological rigor. `prismAId` provides a user-friendly interface that requires no coding skills, enabling researchers from diverse backgrounds to perform comprehensive literature analyses efficiently.
+
+# State of the Field
+AI approaches to support SLRs have been extensively explored, with significant contributions to methodological frameworks and approaches [@van_dinter_automation_2021; @mahadevkar_exploring_2024], as well as advancements in specific tasks like information extraction [@abdullah_systematic_2023]. Building on these foundations, several tools have been developed for scientific applications [@khalil_tools_2022; @affengruber_exploration_2024], including commercial options like @distiller_SR. While these tools demonstrate promise, they often face limitations such as discipline specificity [@liu_autoie_2024], a focus on literature selection rather than extraction [@donoghue_lisc_2019], and the need for substantial human intervention [@kiritchenko_exact_2010; @saldanha_evaluating_2016]. Recent evidence highlights the potential of LLMs to enhance SLRs broadly [@alshami_harnessing_2023] and automate information extraction specifically [@schopow_applications_2023]. Building on this evidence, `prismAId` offers an open-source, community-oriented solution that leverages advanced, state-of-the-art, commercial-grade AI tools for automating information extraction within established SLR protocols, addressing gaps in accessibility, scalability, and methodological rigor.
 
 # Features and Capabilities
 `prismAId` is compatible with any systematic review protocol [@schiavo_prospero_2019], including the widely adopted PRISMA 2020 guidelines [@page_prisma_2021]. It focuses specifically on the analysis phase, which is a key concluding component of all systematic review protocols, alongside the reporting stage.
