### 2025-01-12 01a6d1d
add Cohere command-r7b
