### 2025-01-12 d5beb92
add Cohere command-r7b
