### 2025-01-12 bfa75c1
add Cohere command-r7b
