### 2025-01-12 cee99ff
Update version to 0.6.5
