### 2025-01-12 aeb8770
Update Julia package version to 0.6.5 (#120)
Co-authored-by: GitHub Action <action@github.com>
