### 2025-01-17 034c85f
add fully tested support of DeepSeek chat model, documentation
