+++ model/doc.go
@@ -1,5 +1,5 @@
 // Package model provides functionalities to interact with various large language models (LLMs) such as 
-// OpenAI, Anthropic, Cohere, and Google AI. The package supports querying these LLMs with specified prompts 
+// OpenAI, Anthropic, Cohere, Google AI, and DeepSeek. The package supports querying these LLMs with specified prompts 
 // and retrieving structured results including justifications and summaries. It is designed to offer a unified 
 // interface for integrating different AI models into the application.
 package model
