### 2025-01-18 68d42f0
add fully tested support of DeepSeek chat model, documentation (#121)
