+++ config/configuration.go
@@ -119,6 +119,8 @@ func LoadConfig(tomlConfiguration string, envReader EnvReader) (*Config, error) 
 				llm.ApiKey = envReader.GetEnv("CO_API_KEY")
 			case "Anthropic":
 				llm.ApiKey = envReader.GetEnv("ANTHROPIC_API_KEY")
+			case "DeepSeek":
+				llm.ApiKey = envReader.GetEnv("DEEPSEEK_API_KEY")
 			}
 		}
 
