+++ projects/proj_test.toml
@@ -14,9 +14,9 @@ summary = "no"
 
 [project.llm]
 [project.llm.1]
-provider = "Cohere"
+provider = "DeepSeek"
 api_key = ""
-model = "command-r7b-12-2024"
+model = "deepseek-chat"
 temperature = 0.001
 tpm_limit = 0
 rpm_limit = 0

[…]
