### 2025-01-18 80d2fc4
Update Julia package version to 0.6.6 (#128)
Co-authored-by: GitHub Action <action@github.com>
