### 2025-01-18 fc757a8
add DeepSeek support in text
