+++ paper.md
@@ -44,7 +44,7 @@ Building on the tool's aim, `prismAId` includes the capability to preprocess and
 
 Additionally, `prismAId` integrates with Zotero, a popular open source reference management software, allowing users to access shared collections and groups of scientific literature seamlessly. This integration simplifies document retrieval and enhances collaboration. The connection to Zotero is automatic once the API access is configured, as detailed in the `prismAId` documentation. This feature streamlines the workflow, making it easier for users to manage and analyze large sets of scientific manuscripts.
 
-`prismAId` integrates with various large language models (LLMs) from the four main AI providers: OpenAI, Google Cloud AI, Anthropic, and Cohere. This includes access to state-of-the-art models such as many variants of GPT, Gemini, Claude, and Command. The integration is handled through native APIs, allowing for seamless access to advanced generative AI capabilities. Depending on project needs and available opportunities, support for cloud-based deployments of these models will be expanded. The main results dataset generated by `prismAId` can be exported in two formats: a tabular CSV and a nested JSON structure. These options ensure compatibility with a wide range of analysis tools and software.
+`prismAId` integrates with various large language models (LLMs) from the main AI providers: OpenAI, Google Cloud AI, Anthropic, Cohere, and DeepSeek. This includes access to state-of-the-art models such as many variants of GPT, Gemini, Claude, and Command. The integration is handled through native APIs, allowing for seamless access to advanced generative AI capabilities. Depending on project needs and available opportunities, support for cloud-based deployments of these models will be expanded. The main results dataset generated by `prismAId` can be exported in two formats: a tabular CSV and a nested JSON structure. These options ensure compatibility with a wide range of analysis tools and software.
 
 `prismAId` includes advanced functions designed to support debugging and validation throughout the review process. One key feature is input duplication, which is particularly useful when testing on individual or small sets of manuscripts. This feature facilitates sensitivity analysis, allowing users to experiment with prompt design and other parameters, and evaluate their impact on the results. Another useful capability is the option to generate summaries of manuscripts, enabling manual, ad hoc verification of the extracted outcomes against the main messages of the original papers. Additionally, `prismAId` offers the option to request and store Chain of Thought (COT) justifications. Here, the AI models are prompted to explain their reasoning process and identify the specific parts of the manuscript that support their conclusions, providing greater transparency and accountability.
 
