### 2025-01-18 #7616 @ricboer0
@crvernon, @jhculb, and @philip928lin,
I've done a tiny modification in the paper to align it with models supported in last version of the tool.
