### 2025-01-30 34ae8d5
add Jupyter notebooks instructions to the documentation website
