### 2025-01-31 a5a6cfe
remove user confirmation from review logic and documentation
