### 2025-01-31 b09c71a
129 use from jupyter notebooks (#131)
* add Jupyter notebooks instructions to the documentation website

* update CHANGELOG

* remove user confirmation from review logic and documentation
