### 2025-01-31 a9df747
Update version to 0.6.7 (#133)
Co-authored-by: GitHub Action <action@github.com>
