### 2025-01-31 3ab46be
Update Julia package version to 0.6.7
