### 2025-01-31 fed54f0
rolling back license metadata for Pypi package
