### 2025-02-07 #7616 @ricboer0
@jhculb @crvernon is there anything else I could do to improve the paper?
In the meantime, I have been rebasing the ´joss-submission´ branch with updates from the ´main´ branch. Is this necessary?
