+++ python/pyproject.toml
@@ -17,6 +17,8 @@ classifiers = [
 ]
 requires-python = ">=3.6"
 readme = "README.md"
+license = { text = "AGPL-3.0" }
+license-files = ["LICENSE"]
 
 [project.urls]
 homepage = "https://github.com/open-and-sustainable/prismaid"
