+++ .github/workflows/temp.yml
@@ -129,7 +129,7 @@ jobs:
       - name: Install build tools
         working-directory: ./python
         run: |
-          python -m pip install --upgrade pip setuptools wheel build
+          python -m pip install --upgrade pip hatch wheel build
 
       - name: Build Python package
         working-directory: ./python
