+++ .github/workflows/temp.yml
@@ -129,7 +129,10 @@ jobs:
       - name: Install build tools
         working-directory: ./python
         run: |
-          python -m pip install --upgrade pip hatch hatch-vcs wheel build
+          python -m pip install --upgrade pip hatch wheel build
+      
+      - name: Set Hatch Version
+        run: echo "HATCH_VERSION=$(git describe --tags --always --dirty)" >> $GITHUB_ENV
 
       - name: Build Python package
         working-directory: ./python
