### 2025-02-13 #7616 @crvernon
@ricboer0 - at this point, I just need @jhculb to confirm that their review is complete and that they recommend that your submission is ready for publication.

Also concerning the branch, just let me know which branch you wish to have associated with this thread.  Once all has been completed with the review, you will be asked to conduct a release (whatever your final version number is) and an associated archive (e.g., Zenodo) that is to be named exactly as the paper is titled and having the same author list and order as well as the same license as the code base.  This release and archive will just need to reference whichever branch you have selected.  Let me know which one you one and I'll set it in this thread as the default from which the paper compiles here.
