### 2025-02-14 #7616 @ricboer0
> Also concerning the branch, just let me know which branch you wish to have associated with this thread. Once all has been completed with the review, you will be asked to conduct a release (whatever your final version number is) and an associated archive (e.g., Zenodo) that is to be named exactly as the paper is titled and having the same author list and order as well as the same license as the code base. This release and archive will just need to reference whichever branch you have selected. Let me know which one you one and I'll set it in this thread as the default from which the paper compiles here.

@crvernon I would like to keep this `joss-submission` branch associated. I did setup the connection between this GitHub project releases and Zenodo so it will be quite easy to follow what you just said. Once everything defined I intend to delete the branch because archived and with the only difference of the presence of paper.md and bibliography: could this create any problem to the journal?
