+++ cost/doc.go
@@ -1,7 +1,0 @@
-// Package cost provides utilities for assessing and calculating the costs of processing input prompts 
-// using various models from different providers. It includes functionality for defining model-specific 
-// rates and computing costs based on the number of input tokens.
-//
-// The package is designed to be flexible with support for multiple providers and models, making it 
-// adaptable for different pricing schemes and models that may be integrated in the future.
-package cost
