+++ review.go
@@ -7,15 +7,11 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"github.com/open-and-sustainable/prismaid/check"
 	"github.com/open-and-sustainable/prismaid/config"
 	"github.com/open-and-sustainable/prismaid/convert"
 	"github.com/open-and-sustainable/prismaid/debug"
-	"github.com/open-and-sustainable/prismaid/model"
 	"github.com/open-and-sustainable/prismaid/prompt"
 	"github.com/open-and-sustainable/prismaid/results"
-	"github.com/open-and-sustainable/prismaid/review"
-	"github.com/open-and-sustainable/prismaid/tokens"
 	"github.com/open-and-sustainable/prismaid/zotero"
 	"sync"
 	"time"
