+++ results/save.go
@@ -1,0 +1,196 @@
+package results
+
+import (
+	"encoding/json"
+	"fmt"
+	"path/filepath"
+	"os"
+
+	"github.com/open-and-sustainable/prismaid/config"
+	"github.com/open-and-sustainable/alembica/definitions"
+	"github.com/open-and-sustainable/alembica/utils/logger"
+)
+
+func Save(config *config.Config, results string, filenames []string, keys []string) error {
+	resultsFileName := config.Project.Configuration.ResultsFileName
+	outputFormat := config.Project.Configuration.OutputFormat
+	outputFilePath := resultsFileName + "." + outputFormat
+
+	saveJustificationsAndSummaries(config, resultsFileName, results, filenames)
+
+	if outputFormat == "json" {
+		return saveJSON(outputFilePath, results, filenames)
+	} else if outputFormat == "csv" {
+		return saveCSV(outputFilePath, results, filenames, keys)
+	} else {
+		return fmt.Errorf("unsupported output format: %s", outputFormat)
+	}
+}
+
+func saveJSON(filePath string, resultsString string, filenames []string) error {
+	outputFile, err := os.Create(filePath)
+	if err != nil {
+		logger.Error("Error creating JSON file:")
+		return err
+	}
+	defer outputFile.Close()
+
+	// Start JSON array
+	if err := startJSONArray(outputFile); err != nil {
+		return err
+	}
+
+	// Unmarshal JSON results
+	var parsedResults definitions.Output
+	if err := json.Unmarshal([]byte(resultsString), &parsedResults); err != nil {
+		logger.Error("Error parsing JSON for structured output:", err)
+		return err
+	}
+
+	// Debugging: Check response count
+	fmt.Println("Total JSON responses:", len(parsedResults.Responses))
+
+	// Write each response separately with provider & model metadata
+	for i, response := range parsedResults.Responses {
+		filenameIndex := i % len(filenames) // Prevent index out of range
+		fmt.Println("Processing response", i+1, "/", len(parsedResults.Responses), "Filename:", filenames[filenameIndex])
+
+		modifiedResponse := map[string]interface{}{
+			"provider": response.Provider,
+			"model":    response.Model,
+			"filename": filenames[filenameIndex],
+		}
+
+		// Merge model response into the modified response map
+		var responseData map[string]interface{}
+		if err := json.Unmarshal([]byte(response.ModelResponses[0]), &responseData); err == nil {
+			for key, value := range responseData {
+				modifiedResponse[key] = value
+			}
+		}
+
+		// Convert to JSON string and write it
+		modifiedJSON, err := json.MarshalIndent(modifiedResponse, "", "    ")
+		if err != nil {
+			logger.Error("Error marshaling modified JSON:", err)
+			return err
+		}
+
+		// Ensure newline between JSON objects
+		if i > 0 {
+			if err := writeCommaInJSONArray(outputFile); err != nil {
+				return err
+			}
+		}
+
+		_, err = outputFile.WriteString(string(modifiedJSON) + "\n") // Add newline for readability
+		if err != nil {
+			logger.Error("Error writing JSON to file:", err)
+			return err
+		}
+	}
+
+	// Close JSON array
+	if err := closeJSONArray(outputFile); err != nil {
+		return err
+	}
+
+	logger.Info("JSON results successfully saved to:", filePath)
+	return nil
+}
+
+func saveCSV(filePath string, resultsString string, filenames []string, keys []string) error {
+    outputFile, err := os.Create(filePath)
+    if err != nil {
+        return err
+    }
+    defer outputFile.Close()
+
+    writer := createCSVWriter(outputFile, keys)
+    defer writer.Flush()
+
+    var parsedResults definitions.Output
+    if err := json.Unmarshal([]byte(resultsString), &parsedResults); err != nil {
[…]
