+++ review.go
@@ -1,20 +1,23 @@
 package prismaid
 
 import (
-	"encoding/csv"
-	"github.com/open-and-sustainable/alembica/utils/logger"
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"os"
-	"path/filepath"
+	"strconv"
+	"sync"
+	"time"
+
+	"github.com/open-and-sustainable/alembica/definitions"
+	"github.com/open-and-sustainable/alembica/extraction"
+	"github.com/open-and-sustainable/alembica/utils/logger"
 	"github.com/open-and-sustainable/prismaid/config"
 	"github.com/open-and-sustainable/prismaid/convert"
 	"github.com/open-and-sustainable/prismaid/debug"
 	"github.com/open-and-sustainable/prismaid/prompt"
 	"github.com/open-and-sustainable/prismaid/results"
 	"github.com/open-and-sustainable/prismaid/zotero"
-	"sync"
-	"time"
 )
 
 const (
@@ -37,6 +40,11 @@ func exit(code int) {
 // Global variable to store the timestamps of requests
 var requestTimestamps []time.Time
 var mutex sync.Mutex
+
+// prompts for specific functionalities
+const justification_query = "For each one of the keys and answers you provided, provide a justification for your answer as a chain of thought. In particular, I want a textual description of the few stages of the chin of thought that lead you to the answer you provided and the sentences in the text you analyzes that support your decision. If the value of a key was 'no' or empty '' because of lack of information on that topic in the text analyzed, explicitly report this reason. Please provide only th einformation requested, neither introductory nor concluding remarks."
+const summary_query = "Summarize in very few sentences the text provided to you before for your review."
+
 
 // RunReview is the main function responsible for orchestrating the systematic review process.
 // It takes a TOML string as input, which defines the configuration for the review, and executes 
@@ -129,13 +137,13 @@ func RunReview(tomlConfiguration string) error {
 	if config.Project.Zotero.User != "" {
 		client := &http.Client{}
 		// downlaod pdfs
-		err := zotero.DownloadPDFs(client, config.Project.Zotero.User, config.Project.Zotero.API, config.Project.Zotero.Group, getDirectoryPath(config.Project.Configuration.ResultsFileName))
+		err := zotero.DownloadPDFs(client, config.Project.Zotero.User, config.Project.Zotero.API, config.Project.Zotero.Group, results.GetDirectoryPath(config.Project.Configuration.ResultsFileName))
 		if err != nil {
 			logger.Error("Error:\n%v", err)
 			return err
 		}
 		// convert pdfs
-		err = convert.Convert(getDirectoryPath(config.Project.Configuration.ResultsFileName)+"/zotero", "pdf")
+		err = convert.Convert(results.GetDirectoryPath(config.Project.Configuration.ResultsFileName)+"/zotero", "pdf")
 		if err != nil {
 			logger.Error("Error:\n%v", err)
 			exit(ExitCodeErrorInReviewLogic)
@@ -160,40 +168,26 @@ func RunReview(tomlConfiguration string) error {
 	prompts, filenames := prompt.ParsePrompts(config)
 	logger.Info("Found", len(prompts), "files")
 
-	// build options object
-	options, err := review.NewOptions(config.Project.Configuration.ResultsFileName, config.Project.Configuration.OutputFormat, config.Project.Configuration.CotJustification, config.Project.Configuration.Summary)
+	// convert to JSON format
+	jsonString, err := convertToJSON(config, prompts)
[…]

[…]
