+++ convert/pdf2txt.go
@@ -1,14 +1,16 @@
 package convert
 
 import (
-    "log"
     "os"
     "regexp"
+
     pdf "github.com/ledongthuc/pdf"
 
 	api "github.com/pdfcpu/pdfcpu/pkg/api"
     model "github.com/pdfcpu/pdfcpu/pkg/pdfcpu/model"
     pdfTypes "github.com/pdfcpu/pdfcpu/pkg/pdfcpu/types"
+
+	"github.com/open-and-sustainable/alembica/utils/logger"
 )
 
 // Primary text extraction function using github.com/ledongthuc/pdf
@@ -18,42 +20,42 @@ func readPdf(path string) (string, error) {
     // Open the PDF file
     f, r, err := pdf.Open(path)
     if err != nil {
-        log.Printf("Failed to open PDF: %v", err)
+        logger.Error("Failed to open PDF: %v", err)
         return "", err
     }
     defer f.Close()
 
     totalPage := r.NumPage()
     if totalPage == 0 {
-        log.Println("The PDF contains no pages")
+        logger.Error("The PDF contains no pages")
         return "", nil
     }
 
     for pageIndex := 1; pageIndex <= totalPage; pageIndex++ {
         p := r.Page(pageIndex)
         if p.V.IsNull() {
-            log.Printf("Page %d is null or not available", pageIndex)
+            logger.Error("Page %d is null or not available", pageIndex)
             continue
         }
 
         rows, err := p.GetTextByRow()
         if err != nil {
-            log.Printf("Error retrieving text for page %d: %v", pageIndex, err)
+            logger.Error("Error retrieving text for page %d: %v", pageIndex, err)
             continue
         }
         if len(rows) == 0 {
-            log.Printf("No text rows found on page %d", pageIndex)
+            logger.Error("No text rows found on page %d", pageIndex)
             continue
         }
 
         for _, row := range rows {
             if len(row.Content) == 0 {
-                log.Printf("Empty content on page %d", pageIndex)
+                logger.Error("Empty content on page %d", pageIndex)
                 continue
             }
             line := textsToString(row.Content)
             if line == "" {
-                log.Printf("Converted text is empty on page %d", pageIndex)
+                logger.Error("Converted text is empty on page %d", pageIndex)
             }
             text += line + "\n"
         }
@@ -61,7 +63,7 @@ func readPdf(path string) (string, error) {
 
     // Fallback if no text was extracted
     if text == "" {
-        log.Println("No text extracted from any pages of the PDF, attempting alternative method.")
+        logger.Info("No text extracted from any pages of the PDF, attempting alternative method.")
         return extractTextWithPdfCpu(path)
     }
     return text, nil
@@ -97,7 +99,7 @@ func extractTextWithPdfCpu(filePath string) (string, error) {
 
 	// Optimize the PDF context to fix minor issues
 	if err := api.OptimizeContext(ctx); err != nil {
-		log.Printf("Optimization failed: %v", err)
+		logger.Error("Optimization failed: %v", err)
 	}
 
 	var extractedText string
@@ -105,13 +107,13 @@ func extractTextWithPdfCpu(filePath string) (string, error) {
 	for i := 1; i <= ctx.PageCount; i++ {
 		pageDict, _, _, err := ctx.PageDict(i, false)
 		if err != nil {
-			log.Printf("Error extracting page %d: %v", i, err)
+			logger.Error("Error extracting page %d: %v", i, err)
 			continue
 		}
 
 		contentsEntry, ok := pageDict.Find("Contents")
         if !ok {
-            log.Printf("No content stream found for page %d", i)
+            logger.Error("No content stream found for page %d", i)
             continue
         }
 
[…]
