### 2025-03-16 faaea9f
advance generation and output to take advantage of alembica
