+++ prompt/generator.go
@@ -6,12 +6,35 @@ import (
 	"os"
 	"path/filepath"
 	"sort"
+	"strconv"
 	"strings"
+	"time"
 
 	"github.com/open-and-sustainable/prismaid/config"
+	"github.com/open-and-sustainable/alembica/definitions"
 
 	"github.com/open-and-sustainable/alembica/utils/logger"
 )
+
+// prompts for specific functionalities
+const justification_query = `Provide a JSON object containing justifications for each key and answer provided.
+Format:
+{
+  "justifications": {
+    "<key>": {
+      "reasoning_steps": ["Step 1", "Step 2", "Step 3"],
+      "supporting_sentences": ["Sentence 1", "Sentence 2"]
+    },
+    ...
+  }
+}
+If the answer is 'no' or empty '', include an explanation in the "reasoning_steps" field.`
+
+const summary_query = `Provide a JSON object summarizing the provided text.
+Format:
+{
+  "summary": "Your concise summary here."
+}`
 
 // ParsePrompts reads the configuration and generates a list of prompts along with their corresponding filenames.
 // The function combines different parts of the prompts to create a structured list of inputs.
@@ -23,7 +46,7 @@ import (
 // - Two slices of strings: 
 //   - The first slice contains the generated prompts.
 //   - The second slice contains the filenames associated with each prompt.
-func ParsePrompts(config *config.Config) ([]string, []string) {
+func parsePrompts(config *config.Config) ([]string, []string) {
 	// This slice will store all combined prompts
 	var prompts []string
 	// This slice will store the filenames corresponding to each prompt
@@ -131,4 +154,83 @@ func SortReviewKeysAlphabetically(config *config.Config) []string {
 	}
 	sort.Strings(keys) // Sort keys alphabetically for consistent and logical output
 	return keys
+}
+
+// Obtain input object from TOML config
+func PrepareInput(config *config.Config) (string, []string, error) {
+	prompts, filenames := parsePrompts(config)
+
+	logger.Info("Generating input JSON with %d prompts.", len(prompts))
+
+	// Populate metadata
+	jsonSchema := definitions.Input{
+		Metadata: definitions.InputMetadata{
+			Version:       config.Project.Version,
+			SchemaVersion: "1.0", // Hardcoded since it's not in TOML
+			Timestamp:     time.Now().Format(time.RFC3339),
+		},
+	}
+
+	// Populate models
+	for _, llm := range config.Project.LLM {
+		jsonSchema.Models = append(jsonSchema.Models, definitions.Model{
+			Provider:    llm.Provider,
+			APIKey:      llm.ApiKey,
+			Model:       llm.Model,
+			Temperature: llm.Temperature,
+			TPMLimit:    int(llm.TpmLimit),
+			RPMLimit:    int(llm.RpmLimit),
+		})
+	}
+	logger.Info("Added %d models to input JSON.", len(jsonSchema.Models))
+
+	// Populate prompts
+	for i, promptText := range prompts {
+		sequenceNumber := 1 // Track sequence numbering dynamically
+
+		// Append the main prompt
+		jsonSchema.Prompts = append(jsonSchema.Prompts, definitions.Prompt{
+			PromptContent:  promptText,
+			SequenceID:     strconv.Itoa(i + 1),
+			SequenceNumber: sequenceNumber,
+		})
+
+		// Add justification query if enabled
+		if config.Project.Configuration.CotJustification == "yes" {
+			sequenceNumber++
+			jsonSchema.Prompts = append(jsonSchema.Prompts, definitions.Prompt{
[…]
