+++ paper.bib
@@ -1,0 +1,307 @@
+@article{page_prisma_2021,
+	title = {The {PRISMA} 2020 statement: an updated guideline for reporting systematic reviews},
+	volume = {10},
+	issn = {2046-4053},
+	shorttitle = {The {PRISMA} 2020 statement},
+	url = {https://doi.org/10.1186/s13643-021-01626-4},
+	doi = {10.1186/s13643-021-01626-4},
+	number = {1},
+	urldate = {2022-11-03},
+	journal = {Systematic Reviews},
+	author = {Page, Matthew J. and McKenzie, Joanne E. and Bossuyt, Patrick M. and Boutron, Isabelle and Hoffmann, Tammy C. and Mulrow, Cynthia D. and Shamseer, Larissa and Tetzlaff, Jennifer M. and Akl, Elie A. and Brennan, Sue E. and Chou, Roger and Glanville, Julie and Grimshaw, Jeremy M. and Hróbjartsson, Asbjørn and Lalu, Manoj M. and Li, Tianjing and Loder, Elizabeth W. and Mayo-Wilson, Evan and McDonald, Steve and McGuinness, Luke A. and Stewart, Lesley A. and Thomas, James and Tricco, Andrea C. and Welch, Vivian A. and Whiting, Penny and Moher, David},
+	month = mar,
+	year = {2021},
+	pages = {89},
+}
+
+
+@article{boero_ai-enhanced_2024,
+	title = {An {AI}-{Enhanced} {Systematic} {Review} of {Climate} {Adaptation} {Costs}: {Approaches} and {Advancements}, 2010–2021},
+	volume = {12},
+	copyright = {https://creativecommons.org/licenses/by/4.0/},
+	issn = {2225-1154},
+	shorttitle = {An {AI}-{Enhanced} {Systematic} {Review} of {Climate} {Adaptation} {Costs}},
+	url = {https://www.mdpi.com/2225-1154/12/8/116},
+	doi = {10.3390/cli12080116},
+	abstract = {This study addresses the critical global challenge of climate adaptation by assessing the inadequacies in current methodologies for estimating adaptation costs. Broad assessments reveal a significant investment shortfall in adaptation strategies, highlighting the necessity for precise cost analysis to guide effective policy-making. By employing the PRISMA 2020 protocol and enhancing it with the prismAId tool, this review systematically analyzes the recent evolution of cost assessment methodologies using state-of-the-art generative AI. The AI-enhanced approach facilitates rapid and replicable research extensions. The analysis reveals a significant geographical and sectoral disparity in research on climate adaptation costs, with notable underrepresentation of crucial areas and sectors that are most vulnerable to climate impacts. The study also highlights a predominant reliance on secondary data and a lack of comprehensive uncertainty quantification in economic assessments, suggesting an urgent need for methodological enhancements. It concludes that extending analyses beyond merely verifying that benefits exceed costs is crucial for supporting effective climate adaptation. By assessing the profitability of adaptation investments, it becomes possible to prioritize these investments not only against similar interventions but also across the broader spectrum of public spending.},
+	language = {en},
+	number = {8},
+	urldate = {2024-08-14},
+	journal = {Climate},
+	author = {Boero, Riccardo},
+	month = aug,
+	year = {2024},
+	pages = {116},
+}
+
+@article{schiavo_prospero_2019,
+	title = {{PROSPERO}: {An} {International} {Register} of {Systematic} {Review} {Protocols}},
+	volume = {38},
+	issn = {0276-3869, 1540-9597},
+	shorttitle = {{PROSPERO}},
+	url = {https://www.tandfonline.com/doi/full/10.1080/02763869.2019.1588072},
+	doi = {10.1080/02763869.2019.1588072},
[…]
