### 2025-04-11 617499b
Update macOS Library (#143)
* Add article, bib and figures

* Add action to verify paper output

* Removed: sections Introduction, and Architecture; Added: short ref. to preprint for extended introduction

* Removed: Discussion

* Cleaned: title; Shortened: Features and Capabilities

* Removed: unused references & to Github repo

* Fixed: figures size

* add 'State of the Field' section as requested in review, https://github.com/openjournals/joss-reviews/issues/7616#issuecomment-2565612368

* add DeepSeek support in text

* prepare CITATION.cff for JOSS
