### 2025-04-11 #7616 @ricboer0
👋 @crvernon, I am so excited! Thank you so much to you and the reviewers for the whole process and feedbacks.

https://doi.org/10.5281/zenodo.15197702
I tagged the release 'joss'. In semantic versioning is equivalent to 0.6.7. I hope this solution could work for you as well.
